requirements.txt
fastapi
uvicorn
pillow
requests
pymysql
python-multipart